Removed the folowing comments:
Issue #4753

Unlike file paths, registry paths are not allowed to start or end with a backslash.  Unless run as an administrator (i.e. with elevated privileges), a key that violate this rule results in an ICE03 compiler error. (Registry paths are not pre-evaluated when running with elevated privileges.) Since many users include trailing backslashes in registry paths, we are silently correcting this error here. If there are multiple trailing backslashes, removing the final one, as done here, will purposely still generate the ICE03 error.